# src/api/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Vital-Guard API",
    "description": "Vital-Guard apis"
  },
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/auth/forgot-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Forgot Password",
        "description": "Forgot Password",
        "operationId": "forgot-password",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/HeaderVersion"
          },
          {
            "$ref": "#/parameters/HeaderLanguage"
          },
          {
            "$ref": "#/parameters/HeaderPlatform"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/forgot_password"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Forgot password in system",
            "schema": {
              "$ref": "#/definitions/responsesOk"
            }
          },
          "500": {
            "description": "Unexpected error.",
            "schema": {
              "$ref": "#/definitions/responsesError"
            }
          }
        }
      }
    },
...
    "/auth/reset-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Reset Password",
        "description": "Reset password",
        "operationId": "reset-password",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/HeaderVersion"
          },
          {
            "$ref": "#/parameters/HeaderLanguage"
          },
          {
            "$ref": "#/parameters/HeaderPlatform"
          },
         
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/reset_password"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User profile in system",
            "schema": {
              "$ref": "#/definitions/responsesOk"
            }
          },
          "500": {
            "description": "Unexpected error.",
            "schema": {
              "$ref": "#/definitions/responsesError"
            }
          }
        }
      }
    },
...
    "/users/profile": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Fetch associated profile",
        "description": "Fetch associated profile",
        "operationId": "profile",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/HeaderVersion"
          },
          {
            "$ref": "#/parameters/HeaderLanguage"
          },
          {
            "$ref": "#/parameters/HeaderPlatform"
          },
          {
            "$ref": "#/parameters/HeaderAuth"
          }
        ],
        "responses": {
          "200": {
            "description": "Fetch users profile",
            "schema": {
              "$ref": "#/definitions/responsesOk"
            }
          },
          "500": {
            "description": "Unexpected error.",
            "schema": {
              "$ref": "#/definitions/responsesError"
            }
          }
        }
      }
    },
...
    "/users/password": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Change password",
        "description": "Change passord",
        "operationId": "password",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/HeaderVersion"
          },
          {
            "$ref": "#/parameters/HeaderLanguage"
          },
          {
            "$ref": "#/parameters/HeaderPlatform"
          },
          {
            "$ref": "#/parameters/HeaderAuth"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/change_password"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Change password in system",
            "schema": {
              "$ref": "#/definitions/responsesOk"
            }
          },
          "500": {
            "description": "Unexpected error.",
            "schema": {
              "$ref": "#/definitions/responsesError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "reset_password": {
      "type": "object",
      "required": [
        "email",
        "otp",
        "password",
        "confirm_password"
        
      ],
      "properties": {
        "email": {
          "type": "string",
          "description": "Email",
          "example": "abc@email.com"
        },
        "otp": {
          "type": "string",
          "description": "otp",
          "example": "1234"
        },
        "password": {
          "type": "string",
          "format": "password",
          "example": "1234"
        },
        "confirm_password": {
          "type": "string",
          "format": "password",
          "example": "1234"
        }
      }
    },
    "forgot_password": {
      "type": "object",
      "required": [
        "email"
      ],
      "properties": {
        "email": {
          "type": "string",
          "example": "test@gmail.com"
        }
      }
    },
...
    "change_password": {
      "type": "object",
      "required": [
        "currentPassword",
        "newPassword"
      ],
      "properties": {
        "currentPassword": {
          "type": "string",
          "description": "Current password",
          "format": "password",
          "example": "12345"
        },
        "newPassword": {
          "type": "string",
          "description": "Current password",
          "format": "password",
          "example": "123456"
        }
      }
    },
...
    "responsesOk": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "code": {
          "type": "string"
        },
        "msg": {
          "type": "string"
        },
        "data": {
          "type": "object"
        }
      }
    },
    "responsesError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "code": {
          "type": "string"
        },
        "msg": {
          "type": "string"
        },
        "error": {
          "type": "object"
        }
      }
    },
...
  },
  "parameters": {
    "HeaderVersion": {
      "name": "X-vital-Guard-Version",
      "in": "header",
      "description": "X-vital-Guard-Version",
      "required": true,
      "type": "string",
      "default": "1.0.0"
    },
    "HeaderLanguage": {
      "name": "Accept-Language",
      "in": "header",
      "description": "Accept-Language",
      "required": true,
      "type": "string",
      "default": "en"
    },
    "HeaderPlatform": {
      "name": "X-vital-Guard-Platform",
      "in": "header",
      "description": "android or ios or web",
      "required": true,
      "type": "string",
      "default": "android"
    },
    "HeaderAuth": {
      "name": "Authorization",
      "in": "header",
      "description": "Authorization",
      "required": true,
      "type": "string"
    },
...
  }
}